added admin actions on users: make/remove admin status, delete user

+++ src/controllers/api/userController.ts
@@ -31,6 +31,11 @@ export default class UserController
         const userId = request.params.id;
         const updatedUser = request.body;
 
+        if(userId !== request.user?._id && !request.user?.isAdmin)
+        {
+            return response.status(403).send({message: "Forbidden: Admins only"});
+        }
+
         if (updatedUser.newPassword || updatedUser.confirmPassword) {
             if (!updatedUser.newPassword || !updatedUser.confirmPassword) {
                 return response.status(400).send('Both password fields are required.');
@@ -90,12 +95,15 @@ export default class UserController
     async deleteUser(request: Request, response: Response)
     {
         const userId = request.params.id;
+        if(userId !== request.user?._id && !request.user?.isAdmin)
+        {
+            return response.status(403).send({message: "Forbidden: Admins only"});
+        }
         const deletedUser = await userService.deleteUser(userId);
         if (!deletedUser) response.status(404).send('User not found');
         else
         {
-            const { password, ...deletedUserWithoutPassword } = deletedUser;
-            response.status(200).json(deletedUserWithoutPassword);
+            response.status(200).json({success: true, user: deletedUser});
         }
     }
 
@@ -107,23 +115,6 @@ export default class UserController
         response.status(200).json(filteredUsers);
     }
 
-    async renderUsers(request: Request, response: Response)
-    {
-        try
-        {
-            const users = await userService.readUsers();
-            if(!users) response.status(404).send('Users data not found');
-            response.render('users', {users});
-        }
-        catch(error)
-        {
-            const errorDate = new Date();
-            const errorDateString = errorDate.toLocaleDateString();
-            const errorTimeString = errorDate.toLocaleTimeString();
-            console.error(`[${errorDateString} @ ${errorTimeString}] Error rendering users page: `, error);
-        }
-    }
-
     async getUserPosts(request: Request, response: Response)
     {
         let userId = request.params.userId;
@@ -131,5 +122,14 @@ export default class UserController
         const posts = await postService.getPostsByUserId(userId, page, 10);
         if(!posts) return response.status(404).send("No posts by user");
         return response.status(200).send({message: `Found posts of user with id ${userId}`, posts: posts});
+    }
+
+    async toggleAdmin(request: Request, response: Response)
+    {
+        let userId = request.params.userId;
+        const user = await userService.toggleAdmin(userId);
+        if(!user)
+            return response.status(404).json({message: "User not found", success: false});
+        return response.status(200).json({message: "User's admin status switched.", success: true, user: user});
     }
 }
